Move datepicker to own module
