+++ frontend/src/components/InputField.tsx
@@ -1,12 +1,11 @@
-import ReactDatePicker from 'react-datepicker';
 import React, {
   MutableRefObject, useEffect, useId, useRef,
 } from 'react';
 import {
-  Controller, Control, UseFormRegisterReturn, FieldError,
+  UseFormRegisterReturn, FieldError,
 } from 'react-hook-form';
 
-type InputStates = 'default' | 'error' | 'disabled';
+export type InputStates = 'default' | 'error' | 'disabled';
 
 export interface FieldProps {
   state?: InputStates;
@@ -51,13 +50,13 @@ export interface RHFFieldProps extends Omit<FieldProps, 'registerReturn' | 'valu
   registerReturn: UseFormRegisterReturn<any>;
 }
 
-const fieldBaseClass = 'border-[1px] rounded-ft-normal px-4 py-[13px] text-ft-button font-ft-label '
+export const fieldBaseClass = 'border-[1px] rounded-ft-normal px-4 py-[13px] text-ft-button font-ft-label '
                        + 'placeholder:text-ft-text-300 mb-4';
 
-const fieldInvalidClass = 'invalid:bg-ft-warning-100 invalid:text-ft-warning-300'
+export const fieldInvalidClass = 'invalid:bg-ft-warning-100 invalid:text-ft-warning-300'
                             + 'invalid:border-ft-warning-300';
 
-const fieldStateClasses = {
+export const fieldStateClasses = {
   default: 'border-ft-neutral-200',
   error: 'border-[3px] border-ft-danger-200 text-ft-danger-200',
   disabled: 'border-ft-neutral-200 text-ft-text-300 bg-ft-input-placeholder',
@@ -125,72 +124,4 @@ function InputField({
   );
 }
 
-type DatePickerProps = {
-  control: Control<any>;
-  state?: InputStates;
-  name: React.InputHTMLAttributes<HTMLInputElement>['name'];
-  placeholder?: string;
-  labelText?: string;
-  helperText?: string;
-  timeGranularityMinutes: number;
-
-  // CSS class extensions for the elems
-  labelClassName?: string;
-  inputClassName?: string;
-  helperTextClassName?: string;
-};
-
-function DatePicker({
-  control,
-  state = 'default',
-  name = '',
-  placeholder,
-  labelText,
-  helperText,
-  labelClassName = '',
-  inputClassName = '',
-  helperTextClassName = '',
-  timeGranularityMinutes,
-}: DatePickerProps) {
-  const id = useId();
-
-  return (
-    <div className="flex flex-col items-start flex-wrap">
-      { labelText ? (
-        <label
-          htmlFor={id}
-          className={`font-ft-label mb-1 ${labelClassName}`}
-        >
-          {labelText}
-        </label>
-      ) : null}
-      <Controller
-        control={control}
-        name={name}
-        render={({
-          field: {
-            onChange, onBlur, value,
-          },
-        }) => (
-          <div className="flex flex-col items-start flex-wrap">
-            <ReactDatePicker
-              className={`${fieldBaseClass} ${fieldStateClasses[state]} ${inputClassName}`}
-              dateFormat="dd.MM.yyyy HH:mm"
-              dropdownMode="select"
-              minDate={new Date()}
-              onBlur={onBlur}
-              onChange={onChange}
-              placeholderText={placeholder}
-              selected={value ? new Date(value) : null}
-              shouldCloseOnSelect
-              showTimeSelect
-              timeIntervals={timeGranularityMinutes}
-            />
-          </div>
-        )}
-      />
-      {helperText ? <p className={`text-ft-text-300 -mt-4 ${helperTextClassName}`}>{helperText}</p> : null}
-    </div>
-  );
-}
-export { InputField, DatePicker };
+export default InputField;
